seperate all modules
